# docs/openapi/cloud-server-openapi.yaml
openapi: 3.0.1
info:
  title: cloud-server · 云存储平台 API
  description: |
    ## 统一门户接入（云存储侧）

    本产品接入《云平台统一门户》所需的接口，分为两组：

    | 分组 | 路径 | 鉴权 | 调用方 |
    |---|---|---|---|
    | 门户内网接口 | `/api/internal/**` | 请求头 `X-Internal-Key` | 门户主动调用 |
    | 门户运维接口 | `/api/admin/portal/**` | Bearer JWT + admin 角色 | 运维/联调手工触发 |

    ### 测试前必做：点右上角 Authorize
    1. **internalKey**：填平台侧下发的密钥（对应 `app.portal.internal-key`）。
       未配置时后端会 fail-closed，一律返回 `HTTP 401 {"code":40001}`。
    2. **bearerAuth**：填 `admin / Admin@123` 调 `/api/auth/login` 拿到的 `accessToken`
       （**只填 token 本身，不要带 "Bearer " 前缀**；注意 accessToken 在响应 `data.accessToken`，不是 `data.token`）。

    ### 两个必须知道的约定
    - **`price_cents` / `amount_cents` 单位一律是「分」**，不是元。
    - **`/internal/provision` 的 HTTP 状态码恒为 200**，成败看 body 里的 `code`
      （`0` 成功 / `4003` sku 不存在 / `4004` 库存不足 / `5001` 内部错误）。POST 默认的 201 会让门户判错。
  version: v1
servers:
- url: http://127.0.0.1:3001/api
  description: Generated server url
...
paths:
...
  /internal/provision:
    post:
      tags:
      - 门户内网接口
      summary: ② 开通（幂等，order_item_id 为幂等键）
      description: |
        门户在用户支付成功后调用，为本产品开通实例（云存储＝给该用户增加存储配额）。

        ### 两条铁律（后端已实现，联调时请据此验证）
        1. **`order_item_id` 是幂等键**：有持久化唯一索引，
           同一个 `order_item_id` **只会开通一次**，重复调用直接返回首次结果（含相同 `instance_id`）。
        2. **先落库再建实例**：服务端先写 provision 记录再授予配额，
           否则门户超时重试会开出两个实例。

        ### HTTP 状态码恒为 200，成败看 body 的 `code`
        | code | 场景 | 门户行为 |
        |---|---|---|
        | `0` | 成功（或幂等命中首次结果） | 正常 |
        | `4003` | sku 不存在/已下架 | **不重试**，立即退款 |
        | `4004` | 库存不足 | **不重试**，立即退款 |
        | `5001` | 内部错误（含 `quantity` 越界、开通进行中） | **会重试**（最多 3 次），用尽后自动退款 |

        ### 入参约束
        - `quantity`：**1~1000**（缺省 1）。越界返回 `5001`，不会误授予资源。
        - 单笔授予总量上限 1 PiB。
        - `period_days`：不传则用 SKU 自身的 `period_days`。
      operationId: provision
      requestBody:
        description: 开通请求（字段名 snake_case）
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionRequest"
            examples:
              最小请求:
                description: 最小请求
                value:
                  order_item_id: 50001
                  portal_user_id: 018f3a2b-7c1d-7000-9e21-3f9a2b8c1d00
                  sku_id: storage-500g
                  quantity: 1
                  period_days: 30
              带 extra（原样落库）:
                description: 带 extra（原样落库）
                value:
                  order_item_id: 50002
                  portal_user_id: 018f3a2b-7c1d-7000-9e21-3f9a2b8c1d00
                  sku_id: storage-100g
                  quantity: 2
                  period_days: 90
                  extra:
                    渠道: 小程序
                    orderNo: A-1
        required: true
      responses:
        "200":
          description: HTTP 200 恒定；成功时 body.code=0
          content:
            application/json:
              examples:
                成功:
                  description: 成功
                  value:
                    code: 0
                    message: ok
                    data:
                      instance_id: st-000001
                      expire_at: 2026-11-08T02:42:19Z
                      quantity: 1
        "401":
          description: 内部密钥缺失或错误
          content:
            application/json:
              example:
                code: 40001
                message: 无效的内部调用凭证
                data: null
      security:
      - internalKey: []
...
components:
  schemas:
...
    JsonNode:
      type: object
      description: 门户透传的附加信息，**任意结构**，本侧原样落库到 `portal_provision.extra`
      nullable: true
      example:
        渠道: 小程序
        orderNo: A-1
    ProvisionRequest:
      required:
      - order_item_id
      - portal_user_id
      - sku_id
      type: object
      properties:
        order_item_id:
          type: integer
          description: 门户订单项 id。**幂等键**：同一个值重复调用只会开通一次，直接返回首次结果
          format: int64
          example: 50001
        portal_user_id:
          maxLength: 36
          type: string
          description: 门户用户 UUID（门户 JWT 的 sub）。本产品会据此建立/复用影子本地用户
          example: 018f3a2b-7c1d-7000-9e21-3f9a2b8c1d00
        sku_id:
          maxLength: 64
          type: string
          description: 要开通的 SKU 标识（需存在于商品目录且在售）
          example: storage-500g
        quantity:
          maximum: 1000
          minimum: 1
          type: integer
          description: 数量，**允许范围 1~1000**（缺省 1）。越界返回 `code=5001`，不会误授予资源
          format: int32
          example: 1
          default: 1
        period_days:
          type: integer
          description: 周期天数。不传则用该 SKU 自身的 `period_days`
          format: int32
          example: 30
        extra:
          $ref: "#/components/schemas/JsonNode"
      description: 开通请求（字段名 snake_case，与门户契约 §3.2 逐字对齐）
...
  securitySchemes:
    internalKey:
      type: apiKey
      description: 门户 → 本产品的内部调用密钥（后端配置项 app.portal.internal-key）。仅作用于 /internal/**，未配置该密钥时后端拒绝一切内部调用（fail-closed）。
      name: X-Internal-Key
      in: header
    bearerAuth:
      type: http
      description: 两种令牌都被接受：① 本系统自签 HS256（POST /api/auth/login 获取，管理端接口用这个）；② 门户签发的
        RS256（业务路由用，后端按 JWKS 本地验签）。
      scheme: bearer
      bearerFormat: JWT